+++ crates/task-runner/src/task_runner.rs
@@ -133,6 +133,19 @@ impl<'task> TaskRunner<'task> {
         let is_primary = context.is_primary_target(&self.task.target);
 
         self.report.output_prefix = Some(context.get_target_prefix(&self.task.target));
+        self.report.output_style = if is_primary
+            && !self
+                .app_context
+                .workspace_config
+                .experiments
+                .explicit_task_output_style
+        {
+            // Primary targets always display their output, unless the
+            // experiment is enabled, which honors the configured style
+            None
+        } else {
+            self.task.options.output_style
+        };
         self.report.primary = is_primary;
 
         let result = self.internal_run(context, node).await;